4-DjangoAPI/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-computed-property==0.3.0
djangorestframework==3.12.4
idna==3.3
python-Levenshtein==0.12.2
pytz==2021.3
rapidfuzz==1.8.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
